🐛 fix: keep the numbers 0/1 out of the Bool branch
`normalizeValues` checked `as? Bool` before `as? NSNumber`. The comment
was right that Bool bridges to NSNumber in ObjC, but the guard was wider
than the intent: `NSNumber -> Bool` casting succeeds for exactly 0 and 1,
so the JSON numbers 0/1 (and 0.0/1.0) were swallowed and normalized to
"false"/"true".

Discriminate on the CoreFoundation type id instead. Every value reaching
this method comes from JSONSerialization, which boxes JSON booleans as
__NSCFBoolean and numbers as __NSCFNumber — so CFBooleanGetTypeID is the
only reliable discriminator when a cast cannot tell them apart.

The existing `normalizesBooleanValues` test is the negative control for
the removed branch: dropping the CF check reds it, confirming the
mechanism is covered rather than assumed.

Reachability is low but real — the GBNF grammar constrains every field to
string, so a number only arrives from an unconstrained call or a
generation that breaks the grammar. No user-visible incident is known;
this lands for correctness and cross-language parity.

`normalizeValues` moves to a sibling file: the original sat at the
400-line file_length cap exactly, leaving no room for the why-comment.
The split follows the LlamaCppService+*.swift precedent in this
directory; `nonisolated` on the extension is load-bearing (Pattern 3).

Closes #1150
